implements Admin ADD and setups up EDIT and DELETE
